requirements.txt
selenium >= 4.3.0
robotframework >= 4.1.3
robotframework-pythonlibcore >= 4.4.1
robotframework-seleniumlibrary
robotframework-screencapLibrary
robotframework-selenium2library
robotframework-faker
robotframework-excellib
robotframework-requests >= 0.9.6
robotframework-jsonlibrary
PyYAML
robotframework-databaselibrary
pip install oracledb
robotframework-archivelibrary
unicode